feat: bottomActions 추가
