+++ src/pages/postDetail/postDetail.styles.ts
@@ -45,49 +45,3 @@ export const BookmarkInfo = styled.div`
   letter-spacing: -0.1px;
   margin-top: 8px;
 `;
-
-export const BottomActions = styled.div`
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background: ${colors.gray[50]};
-  padding: 10px 15px;
-  display: flex;
-  gap: 5px;
-  justify-content: center;
-  border-top: 1px solid ${colors.gray[200]};
-`;
-
-export const ActionButton = styled.button`
-  border: none;
-  border-radius: 12px;
-  font-size: 16px;
-  font-weight: 700;
-  line-height: 22px;
-  letter-spacing: -0.16px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
-  height: 58px;
-  transition: all 0.2s ease;
-
-  &:active {
-    transform: scale(0.98);
-  }
-`;
-
-export const SecondaryButton = styled(ActionButton)`
-  background: ${colors.gray[150]};
-  color: ${colors.gray[800]};
-  flex: 1;
-  max-width: 136px;
-`;
-
-export const PrimaryButton = styled(ActionButton)`
-  background: ${colors.green[200]};
-  color: ${colors.gray[800]};
-  flex: 1;
-  max-width: 202px;
-`;
