+++ src/pages/postDetail/postDetail.tsx
@@ -4,10 +4,6 @@ import {
   PostDetailContainer,
   HeaderImage,
   ContentContainer,
-  BottomActions,
-  ActionButton,
-  SecondaryButton,
-  PrimaryButton,
   BookmarkInfo,
 } from './postDetail.styles';
 import PreferenceSection from '../../components/post/postDetail/preferenceSection';
@@ -26,22 +22,13 @@ import DetailSection from '../../components/post/postDetail/detailSection';
 import ContentSection from '../../components/post/postDetail/contentSection';
 import TabBar from '../../components/post/postDetail/tabBar';
 import DividerBar from '../../components/post/postDetail/dividerBar';
+import BottomActions from '../../components/post/postDetail/bottomActions';
 const PostDetail: React.FC = () => {
   const { id } = useParams<{ id: string }>();
   const postData = mockPostData.find((post) => post.id === Number(id)) || mockPostData[0];
   const [activeTab, setActiveTab] = useState<PostCategoryValue>(
     CATEGORY.POST_CONTENT?.value || 'content',
   );
-
-  const handleSupport = () => {
-    // 응원하기 기능
-    console.log('응원하기');
-  };
-
-  const handleChat = () => {
-    // 대화하기 기능
-    console.log('대화하기');
-  };
 
   return (
     <PostDetailContainer>
@@ -91,36 +78,19 @@ const PostDetail: React.FC = () => {
               tools={postData.preferences.tools}
               method={postData.preferences.method}
             />
+            <BookmarkInfo>
+              북마크 {postData.stats.bookmarks} · 조회 {postData.stats.views}
+            </BookmarkInfo>
           </>
         )}
 
         {activeTab === (CATEGORY.POST_DETAIL?.value || 'detail') && (
           <DetailSection category={postData.category || ''} techTools={postData.techTools || []} />
         )}
-        {/* 선호사항 */}
-
-        {/* 북마크 정보 */}
-        <BookmarkInfo>
-          북마크 {postData.stats.bookmarks} · 조회 {postData.stats.views}
-        </BookmarkInfo>
       </ContentContainer>
 
       {/* 하단 액션 버튼 */}
-      <BottomActions>
-        <SecondaryButton onClick={handleSupport}>
-          <svg width="18" height="21" viewBox="0 0 18 21" fill="none">
-            <path
-              d="M9 1L11.5 7.5L18 8L13.5 12L15 19L9 15L3 19L4.5 12L0 8L6.5 7.5L9 1Z"
-              stroke={colors.gray[800]}
-              strokeWidth="1.5"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-          </svg>
-          응원하기
-        </SecondaryButton>
-        <PrimaryButton onClick={handleChat}>대화하기</PrimaryButton>
-      </BottomActions>
+      <BottomActions />
     </PostDetailContainer>
   );
 };
